update: 기존 이미지 삭제 코드 변경..

+++ start.sh
@@ -11,7 +11,7 @@ docker pull $FRONT_ECR_URL
 echo '기존 컨테이너 삭제 후 신규 컨테이너 생성'
 docker ps -a
 
-check_front=$(docker ps -a | grep front)
+check_cont=$(docker ps -a | grep front)
 if [ $(echo $check_front | awk '{print length($0)}') != 0 ]; then
 docker rm -f front
 fi
@@ -20,5 +20,7 @@ docker create -i --name front -p $FRONT_PORT:$FRONT_PORT $FRONT_ECR_URL
 docker start front
 
 echo '기존 이미지 삭제'
-DELETE_TAG=$(docker images | egrep -v "SIZE|$TAG" | awk '{print $2}')
+check_img=$(docker images | egrep -v "SIZE|$TAG" | awk '{print $2}')
+if [ $(echo $check_img | awk '{print length($0)}') != 0 ]; then
 docker image rm "$ACCOUNT_ID.dkr.ecr.$REGION.amazonaws.com/$FRONT_ECR_NAME:$DELETE_TAG"
+fi
